Camera: Block DIS and Timelapse for HFR
DIS and Timelapse is not supported for HFR scenario.
If user'll start recording with timelapse or DIS with HFR
then crash observed from lower layer.

Fix Made: User'll be notified that HFR/HSR selection is not allowed
for TimeLapse and/ DIS mode. After that, HFR selection will be set to
off and the option will be disabled. It'll not be vice versa as
DIS/TimeLapse has higher priority than HFR for their existance in
primary popup rather than being in secondary.

Change-Id: I42ed3d3d20727ddda1b15c21b3ba85c8c41a82ef
CRs-Fixed: 627373, 627362

Conflicts:
	res/values/qcomstrings.xml
[mikeioannina]: removed HSR code